PlantDestructionExample prefab
- Check out PlantDestructionExample prefab under Prefab\Environment for how to use it

Step1: Attach PlantDestruction.cs to parent
Step 2: Attach particles to script, set amount of hits required to destroy it, and if the particles need an offset when spawning. Particle gamobject should preferably have DestroyTimer script on them.
Step 3: Make sure the parent has a BoxCollider2D and that it has isTrigger.